Refactor handlers out of yargs configurations
