+++ src/handlers/index.js
@@ -1,0 +1,4 @@
+export { default as stringHandler } from './string';
+export { default as numberHandler } from './number';
+export { default as arrayHandler } from './array';
+export { default as booleanHandler } from './boolean';
