+++ src/yargs/commands/number.js
@@ -1,4 +1,4 @@
-import { getRandomIntInRange } from '../../helpers';
+import { numberHandler } from '../../handlers';
 
 /**
  * @type {import('yargs').CommandModule}
@@ -26,18 +26,5 @@ export const NumberCommand = {
       },
     ),
 
-  handler: argv => {
-    const { min, max, count } = argv;
-    if (min > max) {
-      console.error('Min must be less than or equal to max'); // eslint-disable-line no-console
-      return;
-    }
-
-    // Construct number strings of the given count
-    const randomNumbers = [];
-    for (let c = 0; c < count; ++c) { randomNumbers.push(getRandomIntInRange({ min, max })); }
-
-    // Print the generated random strings
-    console.log(randomNumbers.join('\n')); // eslint-disable-line no-console
-  },
+  handler: numberHandler,
 };
